Login portal for customer/manager done
Basically just needed to move around a bunch of files/do some renaming/add an extra login page. So, when you are in the store management login portal, it takes you to the store management view and vice versa. If users want to check either view, they need to either have an account or do a guest login. Currently functionality for actually storing logins (past just one session) is not quite there yet